+++ src/app/views/atlas/atlas.component.ts
@@ -8,6 +8,7 @@ import {AtlasService} from './services/atlas.service';
 import {BackendService} from './services/backend.service';
 import {Filter} from '@/app/shared/filter';
 import {IAtlasLayer, ISummaryControl} from '@/app/views/atlas/atlas.types';
+import {LifeIndexMultipleResponses} from '@/app/views/atlas/constants/response.types';
 import {SummaryControlService} from '@/app/views/atlas/services/summary-control.service';
 
 import {BASE_LAYERS, LAYERS, MAP_OPTIONS} from './constants/atlas.const';
@@ -53,7 +54,7 @@ export class AtlasComponent implements OnInit, OnChanges {
 
     ngOnInit(): void {
         this.backendService.lifeIndex$
-            .subscribe(scores => {
+            .subscribe((scores: LifeIndexMultipleResponses) => {
                 if (this.map) {
                     this.scores = this.backendService.reduceLifeIndexes(scores, this.filter.baseFilter.startYear);
                     this.atlasLayers = this.atlasService.prepareLayers(this.map, BASE_LAYERS, this.scores, this.summaryControl);