+++ src/app/views/stats/table/stats-table-screen.types.ts
@@ -1,0 +1,5 @@
+export interface ITableRow {
+    code: string;
+    name: string;
+    [key: string]: number | string;
+}